% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the SUSNEO app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Named options passed through to \code{\link[golem:with_golem_options]{golem::with_golem_options()}}.}
}
\value{
A \link[shiny:shinyApp]{shiny::shinyApp} object (invisibly).
}
\description{
Starts the Shiny application. Additional arguments are passed to
\code{\link[golem:with_golem_options]{golem::with_golem_options()}}.
}
\examples{
if (interactive()) susneoShinyMatt::run_app()
}
