% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susneoShinyMatt-package.R
\docType{package}
\name{susneoShinyMatt-package}
\alias{susneoShinyMatt}
\alias{susneoShinyMatt-package}
\title{susneoShinyMatt: SUSNEO Shiny App (Take-Home)}
\description{
Shiny app scaffold with modules, R6, tests, and CI.
}
\author{
\strong{Maintainer}: Matthew Simonson \email{mattsimonson87@users.noreply.github.com}

}
\keyword{internal}
